Replace nvstrip with meshoptimizer, upgrade VTX indices to 32-bit
- Remove nvtristriplib, vendor meshoptimizer (zeux/meshoptimizer)
- Replace GenerateStrips()+SetListsOnly(true) with meshopt_optimizeVertexCache -- faster on high-poly meshes
- Widen VTX index type from unsigned short to unsigned int (VertexIndexList_t, Strip_t::pIndices, StripGroupHeader_t::pIndex, Vertex_t::origMeshVertID)
- Update all sizeof/offset calculations for 32-bit indices
- Raise MAXSTUDIOVERTS 65536 -> 524288, MAXSTUDIOTRIANGLES 65536 -> 262144

+++ include/studiomdl/optimize.h
@@ -39,7 +39,7 @@ struct Vertex_t
 	unsigned char boneWeightIndex[MAX_NUM_BONES_PER_VERT];
 	unsigned char numBones;
 
-	unsigned short origMeshVertID;
+	unsigned int origMeshVertID;
 
 	// for sw skinned verts, these are indices into the global list of bones
 	// for hw skinned verts, these are hardware bone indices
@@ -104,9 +104,9 @@ struct StripGroupHeader_t
 
 	int numIndices;
 	int indexOffset;
-	inline unsigned short *pIndex( int i ) const 
-	{ 
-		return (unsigned short *)(((byte *)this) + indexOffset) + i; 
+	inline unsigned int *pIndex( int i ) const
+	{
+		return (unsigned int *)(((byte *)this) + indexOffset) + i;
 	};
 
 	int numStrips;
@@ -120,9 +120,9 @@ struct StripGroupHeader_t
 
 	int numTopologyIndices;
 	int topologyOffset;
-	inline unsigned short *pTopologyIndex( int i ) const 
-	{ 
-		return (unsigned short *)(((byte *)this) + topologyOffset) + i; 
+	inline unsigned int *pTopologyIndex( int i ) const
+	{
+		return (unsigned int *)(((byte *)this) + topologyOffset) + i;
 	};
 };
